services/auth-service/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
marshmallow==3.20.1
boto3==1.33.0
